+++ conftool/model/base.py
