+++ conftool/model/config.py
