+++ conftool/model/instance.py
