+++ conftool/model/query.py
