+++ conftool/model/service_schedule.py
